// rps-backend/src/common/enums.ts
export enum ConfigKey {
  SWAGGER_BASE_PATH = 'SWAGGER_BASE_PATH',

  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_PORT = 'DATABASE_PORT',
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_PASS = 'DATABASE_PASS',
  DATABASE_NAME = 'DATABASE_NAME',

  SMTP_HOST = 'SMTP_HOST',
  SMTP_PORT = 'SMTP_PORT',
  SMTP_AUTH_USER = 'SMTP_AUTH_USER',
  SMTP_AUTH_PASS = 'SMTP_AUTH_PASS',

  MAIL_SENDER = 'MAIL_SENDER',

  NODE_URL = 'NODE_URL',
  CONTRACT_ADDRESS = 'CONTRACT_ADDRESS',
  JUDGE_PRIVATE_KEY = 'JUDGE_PRIVATE_KEY',
}

export enum ValidationType {
  EMAIL,
}

export enum ContractEventType {
  GAME_SETUP = 'GameSetup',
  GAME_MATCHED = 'GameMatched',
}